requirements.txt
requests==2.32.3
google-generativeai>=0.8.3 
PyPDF2==3.0.1
nltk==3.8.1
sentence-transformers==2.7.0
chromadb>=0.4.15 
gradio>=4.44.0  
langdetect==1.0.9
pdfplumber==0.9.0
python-dotenv==1.0.0
numpy>=1.21.0
fastapi>=0.104.0 
pydantic>=2.0   
uvicorn>=0.23.2 